feature: added useLazyGetPublicPostsQuery to MappedPosts
