+++ src/pagesComponents/posts/Posts.module.scss
@@ -1,4 +1,4 @@
-/* Posts.module.scss */
+/* MappedPosts.module.scss */
 .grid {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
