+++ src/shared/api/public/publicApi.ts
@@ -46,5 +46,6 @@ export const {
   useGetPublicPostsQuery,
   useGetTotalUsersCountQuery,
   useLazyGetPostByIdQuery,
-  useLazyGetPostCommentsByPostIdQuery
+  useLazyGetPostCommentsByPostIdQuery,
+  useLazyGetPublicPostsQuery
 } = publicApi;
